Fix bad script update
I changed the name of this script and so github wipped out the changes
made in the script: IE everything was red under the old name and green
under the new name.  So this commit is to bring the script back to
before it was updated, and the next commit will be for showing the
change.