Update script to changes the right way
This commit is part of updating you all with the changes I made to this
script but in my first commit with a name change wiped out all the
commit details.  So no here is what happened and here is the original
text of the changes:

1) Changed name of Other Solutions screen to All Solutions [button
Timer 5], 2) Add back in the deleted square library button on
allSolutionsMenu layout [no help screen reference, sorry], 3) fix
library-selecting scripts so they delete libraries from the list that
cannot be found in solutions ActionLog [buttons Timer 4, Timer Edit 4,
Specific Action 4, Dashboard 16] and Library [Library Setup 1], and 4)
change name of two scripts to align respectively with button and window
names.