Improvements to finds, script comments, and ...
So this update is a failure on my part to keep up with documentation as
I fixed problems and added features.  I ended up spending two days
sadly going thru the database design report comparison to see what
changes I had made and forgotten about.  These are them, and now GitHub
is full informed of all changes to the Library and what will be version
3.1 by next month.