Add Script file differences and update ReadMe
Use Script file differences to copy just the script section of the
current FileMaker Database Design Report.  Commit it.  Then paste in
the script section from your updated database design report to see all
changes between current and updated versions.  This a large file and
you will be see the changes only in the GitHub desktop tool; not on the
GitHub website.  Do this comparison  to basically insure everything is
documented in an updated version.