Enhanced unlink discovery of linked records
Before system just checked if a test template was in use, which meant
users could not unlink a test from a test section not using a test that
was in use in a different test section.  Now the script allows users to
unlink tests from test section where they are not use and informs users
of test sections where they are in use.