Failed attempt to create view showing key relationships
Perhaps at some point this screen may be turned into a useful education
tool.  The intention is to help users see the fields used to create
relationships between records in different tables.  Vocabulary is a
problem with this view, as well as color.  NOTE: arrows where added to
screenshot and are not on either layout.