Version 4 update from version 3
This new version is a major overhaul of the application to make it even
easier to use and update.  To see all documented issues, enhancements,
etc. go to the Library/Version 4.0 Issues Checklists folder.  NOTE:
This is part 1 of the update to GitHub focused on just capturing
changes to scripts, and script additions and deletions.  Part 2 will be
an update of script name and location changes.  Part 3 will wrap-up
this update with the DDR (database design report), screen shots of
layouts, etc.