Version 4 Part 3D of update. DONE!
This is just the checklist to insure everything needed to document the
changes from version 3 to 4 has been updated and accounted for.  As
part of the Help update, I will do screenshots to show differences
between the layouts of the two databases, and upload screenshots of the
tables in the graphical relationship editor.