Update 4.0d Part11
This is the final set of changes to the library 4.0d update before the release.  (I hope!  Fingers crossed!)  One minor fix to the QV script to insure user goes to the correct layout when in inventory mode, and a final run of the DDR Report to catch any all changes made to the library up to this date.