Update 4.0d Part12
So maybe this will be the final update before the release?  1) Moved the reference database folder to its own repository as it used by all apps.  2) Removed folders and files that seemed like a good idea but I'm not going to maintain because their info is in the DDR.  3) Changed the name of the update checklist folder to update information (looking for a name that says here is the update information).  4) Changed a couple of scripts and layouts and so upated the script text files and also the DDR so it now reflects these last minute changes to 4.0d.