Update 4.0d Part14
Discovered before this release that the web button on all copyright layouts was linked to a recently updated script which no longer works for opening websites.  So switched scripts and updated the script now used to open copyright websites.  Had to rerun the DDR to capture these changes, which will now be part the of the 4.0d release.